requirements.txt
# 🧠 Core LLM dependencies
transformers==4.48.2
tokenizers==0.21.0
huggingface-hub==0.28.1
safetensors==0.5.2
python-dotenv==1.0.1

# ⚙️ Runtime tools
accelerate==1.6.0
bitsandbytes==0.45.5
psutil==6.1.1
requests==2.32.3
tqdm==4.67.1
packaging==24.2

# 🔬 Numerical / CUDA runtime support
numpy==2.2.4
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
triton==3.1.0

# 🧪 Optional: math/sympy/nvidia libs (if used in your code)
sympy==1.13.1
fsspec==2025.3.2

# 🛠 Dev & Linting (optional but good practice)
pylint==3.3.4
isort==6.0.0
mccabe==0.7.0
